requirements.txt
torch==2.1.0
torch-geometric==2.4.0
transformers==4.35.0
tree-sitter>=0.25.0
faiss-cpu==1.7.4
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
PyYAML==6.0.1
numpy==1.24.3
scikit-learn==1.3.2
tqdm==4.66.1
wandb==0.16.0